2/4/2023 5:51pm
I went from a simple nav bar to a hero image. Worked on search bar, added effect to the career list and added a calculator. I need to move the calculator to the right of the pay period deductions as well as finish the js for the checkbook and finally renovate the footer.